Rename Trial 3 to Trial 3a
Nina's code did not work so we went to the processing web site and found sources related to our problem. We found a code similar to what we needed so we copied and pasted it and changed some variables to match our specifications. Unfortunately even with this code, the problem remained that our animation file was opening as a link (Chromium). 

Processing kept giving us the error: "RuntineExpectation: Could not load movie file aaa.mp4"

https://www.processing.org/reference/libraries/video/Movie.html