Create Final and Working
This is the final code that worked for. However, it would only run on a Mac PC (not Windows or Linux). The code has the mouseX, mouseY commands as the interactive aspect of the project and it is the trigger that we substituted for the sensor. 

Right Mouse Button - Play
Left Mouse Button - Pause